requirements.txt
numpy
opencv-python
pycocotools
pyyaml
tqdm
matplotlib
pandas
seaborn
scipy
tensorboard
torchmetrics
thop
